+++ src/main/java/com/netgrif/application/engine/configuration/ElasticsearchConfiguration.java
@@ -1,9 +1,16 @@
 package com.netgrif.application.engine.configuration;
 
+import com.netgrif.application.engine.configuration.properties.ElasticsearchProperties;
 import com.netgrif.application.engine.configuration.properties.UriProperties;
 import com.netgrif.application.engine.workflow.service.CaseEventHandler;
+import lombok.RequiredArgsConstructor;
 import org.apache.http.HttpHost;
+import org.apache.http.auth.AuthScope;
+import org.apache.http.auth.UsernamePasswordCredentials;
+import org.apache.http.client.CredentialsProvider;
+import org.apache.http.impl.client.BasicCredentialsProvider;
 import org.elasticsearch.client.RestClient;
+import org.elasticsearch.client.RestClientBuilder;
 import org.elasticsearch.client.RestHighLevelClient;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
@@ -12,13 +19,8 @@ import org.springframework.data.elasticsearch.core.ElasticsearchOperations;
 import org.springframework.data.elasticsearch.core.ElasticsearchRestTemplate;
 
 @Configuration
+@RequiredArgsConstructor
 public class ElasticsearchConfiguration {
-
-    @Value("${spring.data.elasticsearch.url}")
-    private String url;
-
-    @Value("${spring.data.elasticsearch.searchport}")
-    private int port;
 
     @Value("${spring.data.elasticsearch.index.petriNet}")
     private String petriNetIndex;
@@ -32,11 +34,9 @@ public class ElasticsearchConfiguration {
     @Value("${spring.data.elasticsearch.reindex}")
     private String cron;
 
-    private final UriProperties uriProperties;
+    private final ElasticsearchProperties elasticsearchProperties;
 
-    public ElasticsearchConfiguration(UriProperties uriProperties) {
-        this.uriProperties = uriProperties;
-    }
+    private final UriProperties uriProperties;
 
     @Bean
     public String springElasticsearchReindex() {
@@ -65,9 +65,18 @@ public class ElasticsearchConfiguration {
 
     @Bean
     public RestHighLevelClient client() {
-
-        return new RestHighLevelClient(
-                RestClient.builder(new HttpHost(url, port, "http")));
+        RestClientBuilder builder = RestClient.builder(new HttpHost(elasticsearchProperties.getUrl(), elasticsearchProperties.getSearchPort()));
+        if (hasCredentials()) {
+            CredentialsProvider credentialsProvider = new BasicCredentialsProvider();
+            credentialsProvider.setCredentials(AuthScope.ANY,
+                    new UsernamePasswordCredentials(
+                            elasticsearchProperties.getUsername(),
+                            elasticsearchProperties.getPassword()
+                    )
+            );
+            builder.setHttpClientConfigCallback(httpClientBuilder -> httpClientBuilder.setDefaultCredentialsProvider(credentialsProvider));
+        }
+        return new RestHighLevelClient(builder);
     }
 
     @Bean
@@ -78,5 +87,10 @@ public class ElasticsearchConfiguration {
     @Bean
     public CaseEventHandler caseEventHandler() {
         return new CaseEventHandler();
+    }
+
+    private boolean hasCredentials() {
+        return elasticsearchProperties.getUsername() != null && !elasticsearchProperties.getUsername().isBlank() &&
+                elasticsearchProperties.getPassword() != null && !elasticsearchProperties.getPassword().isBlank();
     }
 }