+++ src/main/java/com/netgrif/application/engine/elastic/domain/ElasticTask.java
@@ -64,8 +64,6 @@ public class ElasticTask {
 
     private String userId;
 
-    @JsonSerialize(using = LocalDateTimeSerializer.class)
-    @JsonDeserialize(using = LocalDateTimeDeserializer.class)
     @Field(type = FieldType.Date, format = DateFormat.date_hour_minute_second_millis)
     private LocalDateTime startDate;