+++ src/main/java/com/netgrif/application/engine/elastic/serializer/LocalDateTimeJsonSerializer.java
@@ -1,0 +1,18 @@
+package com.netgrif.application.engine.elastic.serializer;
+
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.fasterxml.jackson.databind.SerializerProvider;
+
+import java.io.IOException;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class LocalDateTimeJsonSerializer extends JsonSerializer<LocalDateTime> {
+    private static final DateTimeFormatter FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd'T'HH:mm:ss.SSS");
+
+    @Override
+    public void serialize(LocalDateTime value, JsonGenerator gen, SerializerProvider serializers) throws IOException {
+        gen.writeString(FORMATTER.format(value));
+    }
+}