+++ src/main/java/com/netgrif/application/engine/workflow/service/WorkflowService.java
@@ -98,6 +98,7 @@ public class WorkflowService implements IWorkflowService {
     @Autowired
     protected IInitValueExpressionEvaluator initValueExpressionEvaluator;
 
+    @Lazy
     @Autowired
     protected IElasticCaseMappingService caseMappingService;
 
@@ -110,6 +111,7 @@ public class WorkflowService implements IWorkflowService {
 
     protected IElasticCaseService elasticCaseService;
 
+    @Lazy
     @Autowired
     public void setElasticCaseService(IElasticCaseService elasticCaseService) {
         this.elasticCaseService = elasticCaseService;
@@ -120,6 +122,7 @@ public class WorkflowService implements IWorkflowService {
         if (useCase.getPetriNet() == null) {
             setPetriNet(useCase);
         }
+        checkChangedDataSet(useCase);
         encryptDataSet(useCase);
         useCase = repository.save(useCase);
         try {
@@ -226,7 +229,7 @@ public class WorkflowService implements IWorkflowService {
 
     private void resolveUserRefPermissions(Case useCase, String userListId, Map<String, Boolean> permission) {
         List<String> userIds = getExistingUsers((UserListFieldValue) useCase.getDataSet().get(userListId).getValue());
-        if (userIds != null && userIds.size() != 0) {
+        if (userIds != null && !userIds.isEmpty()) {
             if (permission.containsKey("view") && !permission.get("view")) {
                 useCase.getNegativeViewUsers().addAll(userIds);
             } else {
@@ -239,7 +242,7 @@ public class WorkflowService implements IWorkflowService {
         if (userListValue == null)
             return null;
         return userListValue.getUserValues().stream().map(UserFieldValue::getId)
-                .filter(id -> userService.resolveById(id, false) != null)
+                .filter(id -> userService.existsById(id))
                 .collect(Collectors.toList());
     }
 
@@ -474,7 +477,7 @@ public class WorkflowService implements IWorkflowService {
         useCase.getPetriNet().getDataSet().values().stream().filter(f -> f instanceof TaskField).map(TaskField.class::cast).forEach(field -> {
             if (field.getDefaultValue() != null && !field.getDefaultValue().isEmpty() && useCase.getDataField(field.getStringId()).getValue() != null &&
                     useCase.getDataField(field.getStringId()).getValue().equals(field.getDefaultValue())) {
-                useCase.getDataField(field.getStringId()).setValue(new ArrayList<>());
+                useCase.getDataField(field.getStringId()).setValue(new ArrayList<>(), false);
                 List<TaskPair> taskPairList = useCase.getTasks().stream().filter(t ->
                         (field.getDefaultValue().contains(t.getTransition()))).collect(Collectors.toList());
                 if (!taskPairList.isEmpty()) {
@@ -562,7 +565,7 @@ public class WorkflowService implements IWorkflowService {
             if (value == null)
                 continue;
 
-            dataField.setValue(method.apply(Pair.of(value, encryption)));
+            dataField.setValue(method.apply(Pair.of(value, encryption)), false);
         }
     }
 
@@ -596,4 +599,17 @@ public class WorkflowService implements IWorkflowService {
         }
         return outcome;
     }
+
+    private void checkChangedDataSet(Case useCase) {
+        boolean changed = false;
+        for (DataField data : useCase.getDataSet().values()) {
+            if (data.isChanged()) {
+                changed = true;
+                data.setChanged(false);
+            }
+        };
+        if (changed) {
+            useCase.setLastModifiedDataSet(LocalDateTime.now());
+        }
+    };
 }